+++ content/docs/payments/accept-payments.mdx
@@ -30,9 +30,9 @@ package so you can use it in a node environment.
 
 Here's a sample code on how to use the NodeJS SDK:
 
-<ResCodeBlock options req1='Hide Response' id1='nodeJsSdkSample'/>
+<CodeBlock options req1='Hide Response' id1='nodeJsSdkSample'/>
 
-<ResCodeBlock options req1='Response JSON' id1='nodeJsSdkSampleResponse' className='!bg-white code-war-500' lang='json'/>
+<CodeBlock options req1='Response JSON' id1='nodeJsSdkSampleResponse' className='!bg-white code-war-500' lang='json'/>
 
 
 When a payment is successful, Lazerpay sends a <code>DEPOSIT_TRANSACTION</code> When a payment is successful, Lazerpay sends a
