+++ content/docs/transfers/webhooks.mdx
@@ -31,7 +31,7 @@ information to you as it becomes available, which is a lot more efficient and be
 
 Here are some sample webhook paylods for deposit transactions
 
-<CodeBlock req='' id='webHookPayload' lang='json' className='h-[492px] ' />
+<CodeBlock req1='' id1='webHookPayload' lang='json' className='h-[492px] ' />
 
 <Divider className='w-20 mt-3' />
 
@@ -49,7 +49,7 @@ using the HMAC SHA256 hash algorithm.
 In order to verify this signature came from Lazerpay, you simply have to generate the HMAC 
 SHA256 hash and compare it with the signature received.
 
-<CodeBlock req='' id='webHookPayload2' lang='js' className='h-[492px]' />
+<CodeBlock req1='' id1='webHookPayload2' lang='js' />
 
 <Divider className='w-20 mt-3' mb-4 />
